% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_start}
\alias{run_start}
\title{Start the framework}
\usage{
run_start(
  project = "project",
  envir = "default",
  dref = vector(mode = "character", length = 0),
  tz = Sys.timezone(),
  log = "f_log",
  global_config = file.path(dirname(getwd()), "config", "config.R"),
  local_config = file.path(getwd(), "config.R"),
  f_log = list(file = "log.csv", dir = "d_internal"),
  d_output = c(dir = file.path(dirname(dirname(getwd())), "AnalyticSoftwareOutput")),
  d_input = c(dir = file.path(dirname(dirname(getwd())), "AnalyticSoftwareInput")),
  d_internal = c(dir = file.path(dirname(dirname(getwd())),
    "AnalyticSoftwareInternal"))
)
}
\arguments{
\item{project}{A character vector of length 1 containing the name of the
project.}

\item{envir}{A name for the environment.}

\item{dref}{A character vector with the list of
configuration entries that contain multiple definitions,
each referring to a specific named operating environment.}

\item{tz}{A time zone to use in date and time calculations.}

\item{log}{A name of a variable that contains the definition of the log
location for the project.}

\item{global_config}{A definition of the location of the global
configuration file. Make it NA if you don't want a global configuration
file.}

\item{local_config}{A definition of the location of the local
configuration file. Make it NA if you don't want a global configuration
file.}

\item{f_log}{A definition of the log
location for the project.}

\item{d_output}{A definition of the output
directory.}

\item{d_input}{A definition of the input
directory.}

\item{d_internal}{A definition of the internal
directory}
}
\value{
The configuration environment.
}
\description{
Sets defaults and starts the framework
}
\details{
This is function has no parameters.
}
\examples{
run_start(global_config = NA, local_config = NA)
}
