% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{clean_file_name}
\alias{clean_file_name}
\title{Clean a file name string from illegal characters}
\usage{
clean_file_name(
  inf = "file name",
  suffix = "",
  blacklist = paste0("[^ABCDEFGHIJKLMNOPQRSTUVWXYZ", "abcdefghijklmnopqrstuvwxyz",
    "0123456789", "\\\\s_&\\\\.\\\\-]")
)
}
\arguments{
\item{inf}{A character vector to be cleaned into a file name component.}

\item{suffix}{A suffix to be attached if any, e.g. .PDF or .document.}

\item{blacklist}{A regular expression character class used as a blacklist,
i.e. the negated white list of characters.}
}
\value{
Cleaned file name component.
}
\description{
Make a string used as a file name a legal file name on a file system
by removing illegal characters
}
\details{
File systems tolerate certain characters in a file name. Some characters
are not legal. This function will replace all characters that are not
white listed and replace them with space. Multiple consecutive spaces will
be replaced with a single one and the leading and trailing spaces will
be stripped.
}
\examples{
clean_file_name(c("fred123", "some file", "bad file##", "w##H$A\%t^"))
}
