% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_start_time}
\alias{get_start_time}
\title{Get framework start time}
\usage{
get_start_time()
}
\value{
The time the package was loaded
}
\description{
Get framework start time
}
\examples{
get_start_time()
}
